Überweisung funktioniert jetzt, sowie einzahlen auf sparkonto

+++ Eigene_Bank_DLL_Assembly/Eigene_Bank_DLL_Assembly/IBankManagement.cs
@@ -28,8 +28,8 @@ namespace Eigene_Bank_DLL_Assembly
         void addCreditAccountUser(int _cNumber, int _id);
         int getBankAccountNumber(int _id, int _whichAccount);
         int getAccType(int Accnumber);
-        int getDepositkontostand(int snumber);
-        int getCreditkontostand(int snumber);
+        double getDepositkontostand(int snumber);
+        double getCreditkontostand(int snumber);
         int getDepositAccOwner(int snumber, int whichuser);
         int getCreditAccOwner(int snumber, int whichuser);
